feat: moved expressions to separate crate
