+++ Cargo.toml
@@ -9,6 +9,7 @@ clap = { version = "4.5.9", features = ["derive"] }
 nom = "7.1.3"
 rand = "0.8.5"
 rusty-dice = { path = "crates/rusty-dice" }
+rusty-dice-expressions = { path = "crates/rusty-dice-expressions" }
 
 [workspace]
 resolver = "2"
