+++ crates/rusty-dice-expressions/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rusty-dice-expressions"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+nom = "8.0.0"
+rusty-dice = { path = "../rusty-dice" }
