+++ src/main.rs
@@ -3,8 +3,6 @@ use app::App;
 
 mod app;
 mod cli;
-mod dice;
-mod expressions;
 
 fn main() -> Result<()> {
     let app = App::new();
